requirements.txt
numpy
scipy
pandas
matplotlib
altair
tqdm
filelock
typer
sh
python-docx
spacy-udpipe
cyhunspell
wget
gensim
